+++ src/components/reportes/SextoReporte.module.css
@@ -1,0 +1,292 @@
+.container {
+  font-family: 'Inter', sans-serif;
+  color: #1b1f24;
+  padding: 2rem 5vw;
+  background-color: #f0eeee;
+  padding-bottom: 15vh;
+}
+
+@media (min-width: 1024px) {
+  .container {
+    padding-bottom: 2rem;
+    padding-left: 15vw;
+  }
+}
+
+.header {
+  margin-bottom: 3rem;
+  background-color: #1b1f24;
+  color: #fff;
+  border-radius: 12px;
+  padding: 2rem 5vw;
+
+
+}
+
+.h1 {
+  font-size: 3.5rem;
+  font-weight: 700;
+  margin-bottom: 1rem;
+  color: #99ff66;
+}
+
+.dateText {
+  font-size: 2rem;
+  font-weight: 300;
+  margin-bottom: 0.8rem;
+  color: #9e78b9;
+}
+
+.introText {
+  font-size: 1.3rem;
+  font-weight: 300;
+  margin-bottom: 0.8rem;
+  text-align: justify;
+  padding-right: 5vw;
+  line-height: 1.8;
+  color: #f0eeee;
+}
+
+.section {
+  margin-bottom: 4rem;
+  border-bottom: 1px solid #444;
+  padding-bottom: 2rem;
+}
+
+.section:last-child {
+  border-bottom: none;
+  margin-bottom: 0;
+}
+
+.titlesGroup {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.2rem;
+}
+
+.sectionHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+  margin-bottom: 3rem;
+  flex-wrap: nowrap;
+  background-color: #1b1f24;
+  border-radius: 12px;
+  padding: 2rem 1vw 2rem 5vw;
+
+}
+
+.sectionTitle {
+  flex: 1 1 auto;
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: #9e78b9;
+  margin: 0;
+}
+
+.articleTitle {
+  margin: 0;
+  font-size: 3.5rem;
+  font-weight: 700;
+  color: #99ff66;
+  padding-right: 5vw;
+}
+
+.sectionImageHeader {
+  width: 60%;
+  border-radius: 12px;
+  object-fit: cover;
+  box-shadow: 0 0 40px #99ff6644;
+
+}
+
+.text {
+  max-width: 75vw;
+  margin: auto;
+}
+
+.textSubtitle {
+  font-size: 2rem;
+  font-weight: 300;
+  margin-bottom: 0.8rem;
+  color: #9e78b9;
+  max-width: 75vw;
+  margin: auto;
+}
+
+
+.boxList {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  margin-top: 1rem;
+}
+
+.boxItem {
+  background-color: #1b1f24;
+  border-radius: 12px;
+  padding: 3rem 3rem;
+  display: flex;
+  align-items: flex-start;
+  gap: 1rem;
+  text-align: justify;
+}
+
+.icon {
+  font-size: 1.8rem;
+  line-height: 1;
+  margin-top: 0.2rem;
+  flex-shrink: 0;
+  color: #99ff66;
+}
+
+.boxItem p {
+  margin: 0;
+  font-weight: 300;
+  font-size: 1.15rem;
+  line-height: 1.6;
+  color: #eee;
+  padding-right: 2rem;
+}
+
+@media (max-width: 768px) {
+  .sectionHeader {
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 2rem 2vw 0rem 2vw;
+
+  }
+
+  .sectionImageHeader {
+    width: 100%;
+    max-width: 100%;
+    margin-bottom: 1rem;
+  }
+
+  .sectionTitle {
+    font-size: 1.25rem;
+  }
+
+  .articleTitle {
+    padding-right: 0;
+    font-size: 2.5rem;
+  }
+}
+
+.text {
+  font-size: 1.2rem;
+  font-weight: 300;
+  line-height: 1.8;
+  text-align: justify;
+  margin-bottom: 1.5rem;
+}
+
+
+.firstPartContainer {
+  display: flex;
+  gap: 2rem;
+  margin-bottom: 1.5rem;
+  align-items: flex-start;
+  flex-wrap: nowrap;
+  padding-left: 3vw;
+}
+
+.textContainer {
+  flex: 0 0 60%;
+  max-width: 60%;
+  padding: 0 3vw 0 0;
+}
+
+.videoArea {
+  flex: 0 0 30%;
+  max-width: 30%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: sticky;
+  top: 80px;
+  max-height: calc(100vh - 100px);
+  border-radius: 15px;
+  box-shadow: 0 0 15px #9e78b9;
+  overflow: hidden;
+  padding: 0.5rem;
+  background-color: #1b1f24;
+}
+
+.videoWrapper {
+  position: relative;
+  width: 100%;
+  padding-top: 177.78%;
+  border-radius: 15px;
+  overflow: hidden;
+}
+
+.videoWrapper iframe {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  border: none;
+}
+
+.instagramButton {
+  background-color: #99ff66;
+  color: #121214;
+  font-weight: 700;
+  border: none;
+  border-radius: 20px;
+  padding: 0.4rem 1rem;
+  font-size: 0.9rem;
+  cursor: pointer;
+  text-align: center;
+  text-decoration: none;
+  width: fit-content;
+  margin-top: 1rem;
+  transition: background-color 0.3s ease;
+}
+
+.instagramButton:hover {
+  background-color: #7abc33;
+}
+
+.spotifyWrapper {
+  max-width: 40vw;
+  margin: 2.5rem auto;
+  border-radius: 12px;
+  overflow: hidden;
+}
+
+@media (max-width: 768px) {
+  .firstPartContainer {
+    flex-direction: column;
+  }
+
+  .textContainer,
+  .videoArea {
+    flex: none;
+    max-width: 100%;
+    width: 100%;
+    position: static;
+    max-height: none;
+    box-shadow: none;
+    background-color: transparent;
+    padding: 0;
+    margin-bottom: 1.5rem;
+  }
+
+  .instagramButton {
+    margin-top: 0.5rem;
+  }
+
+  .spotifyWrapper {
+    max-width: 100%;
+    margin: 1.5rem 0;
+  }
+}
+
+.sectionContainer {
+  width: 100%;
+}
